+++ src/topoViewer/backend/topoViewerAdaptorClab.ts
@@ -94,7 +94,10 @@ export class TopoViewerAdaptorClab {
 
 
       var clabName = parsed.name
-
+      var clabPrefix = parsed.prefix;
+      // if (clabPrefix == "") {
+      //   clabPrefix = ""
+      // }
 
 
       // Define the EnvironmentJson object
@@ -104,6 +107,7 @@ export class TopoViewerAdaptorClab {
 
       const environmentJson: EnvironmentJson = {
         workingDirectory: ".",
+        clabPrefix: `${clabPrefix}`,
         clabName: `${clabName}`,
         clabServerAddress: "",
         clabAllowedHostname: hostname,
@@ -112,8 +116,7 @@ export class TopoViewerAdaptorClab {
         deploymentType: "vs-code",
         topoviewerVersion: `${topoViewerVersion}`,
         topviewerPresetLayout: `${this.currentIsPresetLayout.toString()}`,
-        envCyTopoJsonBytes: cytoTopology,
-        envCyTopoJsonBytesAddon: cytoTopology
+        envCyTopoJsonBytes: cytoTopology
       };
 
       // Serialize EnvironmentJson with hyphenated keys
@@ -175,9 +178,9 @@ export class TopoViewerAdaptorClab {
    * @returns An array of Cytoscape elements (`CyElement[]`) representing nodes and edges.
    */
   public clabYamlToCytoscapeElements(yamlContent: string, clabTreeDataToTopoviewer: Record<string, ClabLabTreeNode> | undefined): CyElement[] {
-        const parsed = yaml.load(yamlContent) as ClabTopology;
-      return this.buildCytoscapeElements(parsed, { includeContainerData: true, clabTreeData: clabTreeDataToTopoviewer });
-    }
+    const parsed = yaml.load(yamlContent) as ClabTopology;
+    return this.buildCytoscapeElements(parsed, { includeContainerData: true, clabTreeData: clabTreeDataToTopoviewer });
+  }
 
 
   /**
@@ -193,9 +196,9 @@ export class TopoViewerAdaptorClab {
    * @returns An array of Cytoscape elements (`CyElement[]`) representing nodes and edges.
    */
   public clabYamlToCytoscapeElementsEditor(yamlContent: string): CyElement[] {
-      const parsed = yaml.load(yamlContent) as ClabTopology;
-      return this.buildCytoscapeElements(parsed, { includeContainerData: false });
-    }
+    const parsed = yaml.load(yamlContent) as ClabTopology;
+    return this.buildCytoscapeElements(parsed, { includeContainerData: false });
+  }
 
 
   /**
@@ -248,6 +251,7 @@ export class TopoViewerAdaptorClab {
   private mapEnvironmentJsonToHyphenated(envJson: EnvironmentJson): string {
     const hyphenatedJson = {
       "working-directory": envJson.workingDirectory,
+      "clab-prefix": envJson.clabPrefix,
       "clab-name": envJson.clabName,
       "clab-server-address": envJson.clabServerAddress,
       "clab-allowed-hostname": envJson.clabAllowedHostname,
@@ -256,8 +260,7 @@ export class TopoViewerAdaptorClab {
       "deployment-type": envJson.deploymentType,
       "topoviewer-version": envJson.topoviewerVersion,
       "topoviewer-layout-preset": envJson.topviewerPresetLayout,
-      "EnvCyTopoJsonBytes": envJson.envCyTopoJsonBytes,
-      "EnvCyTopoJsonBytesAddon": envJson.envCyTopoJsonBytesAddon
+      "EnvCyTopoJsonBytes": envJson.envCyTopoJsonBytes
     };
 
     return JSON.stringify(hyphenatedJson, null, 2);
@@ -284,6 +287,8 @@ export class TopoViewerAdaptorClab {
     log.info(`######### status preset layout: ${this.currentIsPresetLayout}`);
 
     const clabName = parsed.name;
+
+
     const parentMap = new Map<string, string | undefined>();
     let nodeIndex = 0;
 
@@ -324,6 +329,7 @@ export class TopoViewerAdaptorClab {
               image: nodeObj.image ?? '',
               index: nodeIndex.toString(),
               kind: nodeObj.kind ?? '',
+              type: nodeObj.type ?? '',
               labdir: `clab-${clabName}/`,
               labels: nodeObj.labels ?? {},
               longname: `clab-${clabName}-${nodeName}`,