+++ src/topoViewerEditor/webview-ui/topoViewerEditorEngine.ts
@@ -4,6 +4,7 @@ import cytoscape from 'cytoscape';
 import edgehandles from 'cytoscape-edgehandles';
 import cola from 'cytoscape-cola';
 import gridGuide from 'cytoscape-grid-guide';
+import leaflet from 'cytoscape-leaf';
 // Import and register context-menu plugin
 import cxtmenu from 'cytoscape-cxtmenu';
 // import 'cytoscape-cxtmenu/cytoscape-cxtmenu.css';
@@ -11,9 +12,14 @@ import cxtmenu from 'cytoscape-cxtmenu';
 import loadCytoStyle from './managerCytoscapeStyle';
 import { VscodeMessageSender } from './managerVscodeWebview';
 import { fetchAndLoadData, fetchAndLoadDataEnvironment } from './managerCytoscapeFetchAndLoad';
-import { ManagerViewportButtons } from './managerViewportButtons';
+import { ManagerSaveTopo } from './managerSaveTopo';
+import { ManagerZoomToFit } from './managerZoomToFit';
+import { ManagerLabelEndpoint } from './managerLabelEndpoint';
+import { ManagerReloadTopo } from './managerReloadTopo';
+import { ManagerAddContainerlabNode } from './managerAddContainerlabNode';
 import { ManagerViewportPanels } from './managerViewportPanels';
 import { ManagerGroupManager } from './managerGroupManager';
+import { ManagerLayoutAlgo } from './managerLayoutAlgo';
 
 
 
@@ -22,6 +28,8 @@ cytoscape.use(edgehandles);
 cytoscape.use(cola);
 cytoscape.use(gridGuide);
 cytoscape.use(cxtmenu);
+cytoscape.use(leaflet);
+
 
 
 /**
@@ -42,8 +50,9 @@ export interface NodeData {
   };
   extraData?: {
     kind?: string;
-    longname?: string;
     image?: string;
+    type?: string
+    longname?: string;
     mgmtIpv4Address?: string;
   };
 }
@@ -73,10 +82,15 @@ class TopoViewerEditorEngine {
   private isViewportDrawerClabEditorChecked: boolean = true; // Editor mode flag
 
   private messageSender: VscodeMessageSender;
-  private viewportButtons: ManagerViewportButtons;
+  private saveManager: ManagerSaveTopo;
+  private zoomToFitManager: ManagerZoomToFit;
+  private labelEndpointManager: ManagerLabelEndpoint;
+  private reloadTopoManager: ManagerReloadTopo;
+  private addNodeManager: ManagerAddContainerlabNode;
   private viewportPanels: ManagerViewportPanels;
   private groupManager: ManagerGroupManager = new ManagerGroupManager();
-
+  /** Layout manager instance accessible by other components */
+  public layoutAlgoManager: ManagerLayoutAlgo = new ManagerLayoutAlgo();
 
 
 
@@ -96,7 +110,7 @@ class TopoViewerEditorEngine {
         return;
       }
       const suppressNotification = true;
-      await this.viewportButtons.viewportButtonsSaveTopo(this.cy, suppressNotification);
+      await this.saveManager.viewportButtonsSaveTopo(this.cy, suppressNotification);
     }, 500); // Wait 500ms after last change before saving
 
     // Listen for topology changes
@@ -192,10 +206,28 @@ class TopoViewerEditorEngine {
     this.initializeEdgehandles();
     this.initializeContextMenu();
 
-    // Initiate viewport buttons and panels
-    this.viewportButtons = new ManagerViewportButtons(this.messageSender);
-    this.viewportPanels = new ManagerViewportPanels(this.viewportButtons, this.cy, this.messageSender);
+    // Initiate managers and panels
+    this.saveManager = new ManagerSaveTopo(this.messageSender);
+    this.zoomToFitManager = new ManagerZoomToFit();
+    this.labelEndpointManager = new ManagerLabelEndpoint();
+    this.reloadTopoManager = new ManagerReloadTopo(this.messageSender);
+    this.addNodeManager = new ManagerAddContainerlabNode();
+    this.viewportPanels = new ManagerViewportPanels(this.saveManager, this.cy, this.messageSender);
     this.groupManager = new ManagerGroupManager();
+    this.layoutAlgoManager = new ManagerLayoutAlgo();
+
+    // Expose layout functions globally for HTML event handlers
+    (window as any).viewportButtonsLayoutAlgo = this.layoutAlgoManager.viewportButtonsLayoutAlgo.bind(this.layoutAlgoManager);
+    (window as any).layoutAlgoChange = this.layoutAlgoManager.layoutAlgoChange.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerLayoutGeoMap = this.layoutAlgoManager.viewportDrawerLayoutGeoMap.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerDisableGeoMap = this.layoutAlgoManager.viewportDrawerDisableGeoMap.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerLayoutForceDirected = this.layoutAlgoManager.viewportDrawerLayoutForceDirected.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerLayoutForceDirectedRadial = this.layoutAlgoManager.viewportDrawerLayoutForceDirectedRadial.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerLayoutVertical = this.layoutAlgoManager.viewportDrawerLayoutVertical.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerLayoutHorizontal = this.layoutAlgoManager.viewportDrawerLayoutHorizontal.bind(this.layoutAlgoManager);
+    (window as any).viewportDrawerPreset = this.layoutAlgoManager.viewportDrawerPreset.bind(this.layoutAlgoManager);
+    (window as any).viewportButtonsGeoMapPan = this.layoutAlgoManager.viewportButtonsGeoMapPan.bind(this.layoutAlgoManager);
+    (window as any).viewportButtonsGeoMapEdit = this.layoutAlgoManager.viewportButtonsGeoMapEdit.bind(this.layoutAlgoManager);
 
     this.setupAutoSave();
 
@@ -331,10 +363,8 @@ class TopoViewerEditorEngine {
             // this.viewportPanels.panelNodeEditor(ele);
             if (ele.data("topoViewerRole") == "dummyChild") {
               console.info("Editing parent of dummyChild: ", ele.parent().first().id());
-              // this.viewportButtons.viewportButtonsPanelGroupManager.panelGroupTogle(ele.parent().first().id());
               this.groupManager.panelGroupToggle(ele.parent().first().id());
             } else if (ele.data("topoViewerRole") == "group") {
-              // this.viewportButtons.viewportButtonsPanelGroupManager.panelGroupTogle(ele.id());
               this.groupManager.panelGroupToggle(ele.id());
             }
           }
@@ -346,7 +376,6 @@ class TopoViewerEditorEngine {
                       <span>Delete Group</span>
                     </div>`,
           select: () => {
-            //this.viewportButtons.viewportButtonsPanelGroupManager.nodeParentRemoval(this.cy);
             this.groupManager.nodeParentRemoval(this.cy);
           }
         }
@@ -432,7 +461,7 @@ class TopoViewerEditorEngine {
       const mouseEvent = event.originalEvent as MouseEvent;
       if (event.target === this.cy && mouseEvent.shiftKey && this.isViewportDrawerClabEditorChecked) {
         console.log("Canvas clicked with Shift key - adding node.");
-        this.viewportButtons.viewportButtonsAddContainerlabNode(this.cy, this.cyEvent as cytoscape.EventObject);
+        this.addNodeManager.viewportButtonsAddContainerlabNode(this.cy, this.cyEvent as cytoscape.EventObject);
       }
     });
 
@@ -467,13 +496,13 @@ class TopoViewerEditorEngine {
           break;
 
         // case (node.data("topoViewerRole") == "dummyChild"):
-        //   console.info("Editing parent of dummyChiled: ", node.parent().id());
-        //   this.viewportButtons.viewportButtonsPanelGroupManager.panelGroupTogle(node.parent().id());
+        //   console.info("Editing parent of dummyChild: ", node.parent().id());
+        //   this.groupManager.panelGroupToggle(node.parent().id());
         //   break;
         // // If the node is a parent, open the panel for that parent.
         // case node.isParent():
         //   console.info("Editing existing parent node: ", node.id());
-        //   this.viewportButtons.viewportButtonsPanelGroupManager.panelGroupTogle(node.id());
+        //   this.groupManager.panelGroupToggle(node.id());
         //   break;
         default:
           break;